Fix Signin and signup integration
